+++ src/main/java/org/greypowergaeservices/services/ProductService.java
@@ -4,5 +4,5 @@ import org.greypowergaeservices.entities.Product;
 
 public interface ProductService {
 
-	public Long createProduct(Product product);
+	public int createProduct(Product product);
 }
